Define an error handler for ForestLiana that allows exceptions to be bubbled outside.
There are many cases where applications using ForestLiana will have some form of
exception reporting or handling such as Sentry. These are used to track exceptions,
collate and report them for handling - especially in production situations.

However a large amount of the controllers within ForestLiana swallow exceptions
and only write log messages. This can be hard to trace and act on properly,
without capturing the log messages and reporting this afterwards. However
this loses a large amount of metadata and stacktrace data.

Instead we suggest adding an error handler that is called in all places
ForestLiana swallows exceptions to give the outer application to handle
exceptions however they want to.