Add initialisation of configuration directories
We had hardcoded the directory used for configuring more data models.
Hardcoding this value made it difficult to set a custom directory to
define these files. It also caused `zeitwerk` to throw a warning when
trying to autoload the extra classes. We added a `config_dirs` accessor
to the configuration module. Adding this attribute allows us to set this
directory ourselves.